Refactor: * Rename braille.PATTERNS_TABLE so it better reflects its intended purpose. * Move code related to braille input into the brailleInput module.
